refactor: seperate different module (edit directory structure)
